v1.0.0 - Loader Avatar info
